Add files via upload
First version of Population Density Data API. A synchronous approach has been chosen with a single endpoint returning the estimated future  population density in a given area. 
 
Important design choices:
 
If the information available for a grid cell in a given time interval is not sufficient to ensure anonymisation of the data, the response will include the value 'LOW_DENSITY' instead of specific population data.
 
Responses may contain two grid cell formats: boundary cells (BOUNDARY_CELL) that specify the vertices that form a rectangle, or Geohash cells (GEOHASH_CELL) that use the Geohash system (https://en.wikipedia.org/wiki/Geohash).
 
Initially this API is intended to be a synchronous API but it is possible that if required, it could evolve in the future.